Moves DefaultCertDruation to ./api/util/duration.go
