+++ pkg/api/util/BUILD.bazel
@@ -4,6 +4,7 @@ go_library(
     name = "go_default_library",
     srcs = [
         "conditions.go",
+        "duration.go",
         "issuers.go",
     ],
     importpath = "github.com/jetstack/cert-manager/pkg/api/util",
