+++ pkg/controller/certificaterequests/venafi/BUILD.bazel
@@ -14,7 +14,6 @@ go_library(
         "//pkg/internal/venafi:go_default_library",
         "//pkg/issuer:go_default_library",
         "//pkg/logs:go_default_library",
-        "//pkg/util/api:go_default_library",
         "//vendor/github.com/Venafi/vcert/pkg/endpoint:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/api/errors:go_default_library",
         "//vendor/k8s.io/client-go/listers/core/v1:go_default_library",
